requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
Pillow==10.0.0
proglog==0.1.10
pydub==0.25.1
python-dotenv==1.0.0
python-telegram-bot==20.3
regex==2023.6.3
requests==2.31.0
sniffio==1.3.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
